[Rename] Dto, Exception 폴더로 구분
