+++ src/main/java/programo/_pro/dto/authDto/DupCheckDto.java
@@ -1,4 +1,4 @@
-package programo._pro.dto;
+package programo._pro.dto.authDto;
 
 import lombok.Getter;
