+++ src/main/java/programo/_pro/dto/teamDto/TeamCardDto.java
@@ -1,4 +1,4 @@
-package programo._pro.dto;
+package programo._pro.dto.teamDto;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
