+++ src/main/java/programo/_pro/dto/userDto/CustomUserDetails.java
@@ -1,4 +1,4 @@
-package programo._pro.dto;
+package programo._pro.dto.userDto;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
