+++ src/main/java/programo/_pro/dto/userDto/UserInfo.java
@@ -1,4 +1,4 @@
-package programo._pro.dto;
+package programo._pro.dto.userDto;
 
 import programo._pro.entity.User;
 import lombok.*;
