+++ src/main/java/programo/_pro/global/exception/teamException/AlreadyJoinedTeamException.java
@@ -1,4 +1,4 @@
-package programo._pro.global.exception;
+package programo._pro.global.exception.teamException;
 
 public class AlreadyJoinedTeamException extends RuntimeException {
     public AlreadyJoinedTeamException() {
