+++ src/main/java/programo/_pro/global/exception/teamException/NotFoundTeamException.java
@@ -1,4 +1,4 @@
-package programo._pro.global.exception;
+package programo._pro.global.exception.teamException;
 
 public class NotFoundTeamException extends RuntimeException {
     public NotFoundTeamException() {
