+++ src/main/java/programo/_pro/service/ChatService.java
@@ -6,11 +6,11 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
 import programo._pro.dto.ChatMessageRequest;
-import programo._pro.dto.ChatMessageResponse;
+import programo._pro.dto.chatDto.ChatMessageResponse;
 import programo._pro.entity.ChatRoom;
 import programo._pro.entity.Message;
 import programo._pro.entity.User;
-import programo._pro.global.exception.NotFoundChatException;
+import programo._pro.global.exception.chatException.NotFoundChatException;
 import programo._pro.repository.ChatRoomRepository;
 import programo._pro.repository.MessageRepository;
 import programo._pro.repository.UserRepository;
