+++ src/main/java/programo/_pro/service/ChatbotService.java
@@ -5,8 +5,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 import org.springframework.stereotype.Service;
 import programo._pro.entity.*;
-import programo._pro.global.exception.NotFoundChatException;
-import programo._pro.global.exception.NotFoundTeamException;
+import programo._pro.global.exception.chatException.NotFoundChatException;
+import programo._pro.global.exception.teamException.NotFoundTeamException;
 import programo._pro.repository.*;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
