+++ src/main/java/programo/_pro/service/TeamService.java
@@ -3,17 +3,17 @@ package programo._pro.service;
 import org.springframework.transaction.annotation.Transactional;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import programo._pro.dto.TeamCardDto;
-import programo._pro.dto.TeamCreateRequest;
-import programo._pro.dto.TeamMainDto;
-import programo._pro.dto.TeamMemberDto;
+import programo._pro.dto.teamDto.TeamCardDto;
+import programo._pro.dto.teamDto.TeamCreateRequest;
+import programo._pro.dto.teamDto.TeamMainDto;
+import programo._pro.dto.teamDto.TeamMemberDto;
 import programo._pro.entity.Team;
 import programo._pro.entity.TeamMember;
 import programo._pro.entity.User;
-import programo._pro.global.exception.AlreadyJoinedTeamException;
-import programo._pro.global.exception.NotFoundTeamException;
-import programo._pro.global.exception.NotFoundUserException;
-import programo._pro.global.exception.NotJoinedTeamException;
+import programo._pro.global.exception.teamException.AlreadyJoinedTeamException;
+import programo._pro.global.exception.teamException.NotFoundTeamException;
+import programo._pro.global.exception.userException.NotFoundUserException;
+import programo._pro.global.exception.teamException.NotJoinedTeamException;
 import programo._pro.repository.TeamRepository;
 import programo._pro.repository.TeamMemberRepository;
 import programo._pro.repository.UserRepository;
