+++ src/main/java/programo/_pro/service/UserService.java
@@ -1,9 +1,9 @@
 package programo._pro.service;
 
 import programo._pro.dto.AuthenticationToken;
-import programo._pro.dto.UserDto;
+import programo._pro.dto.userDto.UserDto;
 import programo._pro.entity.User;
-import programo._pro.global.exception.NotFoundUserException;
+import programo._pro.global.exception.userException.NotFoundUserException;
 import programo._pro.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -11,8 +11,6 @@ import org.springframework.security.core.userdetails.UserDetailsService;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
