+++ src/test/java/programo/_pro/controller/AuthControllerTest.java
@@ -10,11 +10,9 @@ import org.mockito.Mock;
 import org.mockito.junit.jupiter.MockitoExtension;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.MvcResult;
 import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import programo._pro.dto.SignInDto;
-import programo._pro.dto.SignUpDto;
-import programo._pro.dto.UserInfo;
+import programo._pro.dto.authDto.SignUpDto;
+import programo._pro.dto.userDto.UserInfo;
 import programo._pro.entity.User;
 import programo._pro.service.AuthService;
